requirements.txt
# 기본 의존성
unstructured[pdf]>=0.10.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
streamlit>=1.28.0
plotly>=5.15.0
Pillow>=9.0.0
openpyxl>=3.0.0

# 향상된 파싱을 위한 의존성
layoutparser>=0.3.4
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
scikit-image>=0.21.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0

# 추가 유틸리티
pathlib2>=2.3.0
python-multipart>=0.0.6 